requirements.txt
streamlit>=1.36.0
streamlit-folium>=0.20.0
folium>=0.17.0
geopandas>=0.14.0
shapely>=2.0.0
anthropic>=0.39.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
pandas>=2.0.0
python-dotenv>=1.0.0
